Extract optimize manifests logic
